Enable adapter before using it
This operation is fast if the adapter is already enabled
because no change occurs. If the adapter is off or the
adapter is blocked it enables the adapter or unblocks and
enables the adapter.

Right now unblocking is handled by shelling out to rfkill
but it would be cool if we could implement that functionality
using a library (not pyric probably) or doing it ourselves.